+++ init/52_copy_gsearch_war_to_webapps.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
-mv ${TMP_WORKDIR}/fedoragsearch-2.8/fedoragsearch.war ${FEDORA_HOME}/tomcat/webapps/
+if [ -f /tmp/FEDORA_NEEDS_INSTALL ];
+then
+  mv ${TMP_WORKDIR}/fedoragsearch-2.8/fedoragsearch.war ${FEDORA_HOME}/tomcat/webapps/
+fi
