+++ init/57_stop_tomcat.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
-killall -9 java
+if [ -f /tmp/FEDORA_NEEDS_INSTALL ];
+then
+  killall -9 java
+fi
