+++ init/70_deploy_dgi_gsearch_extensions.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
-cp ${TMP_WORKDIR}/gsearch_extensions-${GSEARCH_DGI_EXTENSION_VERSION}-jar-with-dependencies.jar ${FEDORA_HOME}/tomcat/webapps/fedoragsearch/WEB-INF/lib
+if [ -f /tmp/FEDORA_NEEDS_INSTALL ];
+then
+  cp ${TMP_WORKDIR}/gsearch_extensions-${GSEARCH_DGI_EXTENSION_VERSION}-jar-with-dependencies.jar ${FEDORA_HOME}/tomcat/webapps/fedoragsearch/WEB-INF/lib
+fi
