+++ init/76_add_additional_dgi_gsearch_xslts.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
 
-cp ${TMP_WORKDIR}/basic-solr-config/foxmlToSolr.xslt ${FEDORA_HOME}/tomcat/webapps/fedoragsearch/WEB-INF/classes/fgsconfigFinal/index/FgsIndex/
-cp -r ${TMP_WORKDIR}/basic-solr-config/islandora_transforms ${FEDORA_HOME}/tomcat/webapps/fedoragsearch/WEB-INF/classes/fgsconfigFinal/index/FgsIndex/
+if [ -f /tmp/FEDORA_NEEDS_INSTALL ];
+then
+  cp ${TMP_WORKDIR}/basic-solr-config/foxmlToSolr.xslt ${FEDORA_HOME}/tomcat/webapps/fedoragsearch/WEB-INF/classes/fgsconfigFinal/index/FgsIndex/
+  cp -r ${TMP_WORKDIR}/basic-solr-config/islandora_transforms ${FEDORA_HOME}/tomcat/webapps/fedoragsearch/WEB-INF/classes/fgsconfigFinal/index/FgsIndex/
+fi
