+++ init/80_deploy_solr_xml_to_catalina.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
-cp ${TMP_WORKDIR}/solr.xml ${FEDORA_HOME}/tomcat/conf/Catalina/localhost/
+if [ -f /tmp/FEDORA_NEEDS_INSTALL ];
+then
+  cp ${TMP_WORKDIR}/solr.xml ${FEDORA_HOME}/tomcat/conf/Catalina/localhost/
+fi
